linux: add import/export function for TLS pre-shared keys
The existing import/export function do not handle different version of
the interchange format nor do the handle the HMAC independent of the
version. Thus allow the caller to select version and HMAC independently
when exporting resp. importing. This makes this interface also future
proof when new HMAC or key lengths are added to the spec.

The pre-shared key interchange format also has 'no transform' option
when the configured key should be used as retained key.

Signed-off-by: Daniel Wagner <dwagner@suse.de>