linux: export keys to config
The recent change where key import function out of the JSON parser to
the connect path, dropped the feature to export the keys back to the
config.

The keys are only necessary to be loaded back to the user memory when
the configuration is dumped. Thus hook the export functionality into the
dump code path.

Signed-off-by: Daniel Wagner <dwagner@suse.de>